requirements.txt
Flask
Flask-SocketIO
websocket-client
gunicorn